Separate assume-related code into assume.go and assume_test.go
